refactor: TC-51 Move utility files into utility
