+++ apps/api/src/main/java/dk/treecreate/api/shipmondo/utility/DangerousGoods.java
@@ -1,7 +1,6 @@
-package dk.treecreate.api.shipmondo;
+package dk.treecreate.api.shipmondo.utility;
 
 public class DangerousGoods {
-    
    
     private String ard_class;
     private String un_number;
