+++ apps/api/src/main/java/dk/treecreate/api/shipmondo/utility/Receiver.java
@@ -1,0 +1,6 @@
+package dk.treecreate.api.shipmondo.utility;
+
+public class Receiver {
+        
+    private Address address;
+}
