+++ apps/api/src/main/java/dk/treecreate/api/shipmondo/utility/Sender.java
@@ -1,0 +1,6 @@
+package dk.treecreate.api.shipmondo.utility;
+
+public class Sender {
+   
+    private Address address;    
+}
